refactor(X2-type): harden type safety in LLM client + extract integrity floor
Consolidate LLM client request/response typing and graph-extract schema validation. X2-type (safety-hardened reflect entry + schema-bound extract) improves type safety, reduces runtime failures, and establishes a baseline for future integrity work.
