+++ src/commands/improve/extract.ts
@@ -29,7 +29,7 @@ import fs from "node:fs";
 import path from "node:path";
 import { assembleAsset } from "../../core/asset/asset-serialize";
 import { resolveStashDir, timestampForFilename } from "../../core/common";
-import type { AkmConfig, LlmConnectionConfig } from "../../core/config/config";
+import type { AkmConfig, LlmProfileConfig } from "../../core/config/config";
 import { getDefaultLlmConfig, loadConfig } from "../../core/config/config";
 import { ConfigError, UsageError } from "../../core/errors";
 import { appendEvent } from "../../core/events";
@@ -207,7 +207,7 @@ export interface AkmExtractOptions {
    * Override the LLM chat function (test seam). Defaults to {@link chatCompletion}.
    */
   chat?: (
-    config: LlmConnectionConfig & { supportsJsonSchema?: boolean },
+    config: LlmProfileConfig,
     messages: ChatMessage[],
     options?: { timeoutMs?: number; responseSchema?: Record<string, unknown> },
   ) => Promise<string>;
@@ -444,7 +444,7 @@ async function processSession(
   sessionRef: SessionRef,
   stashDir: string,
   config: AkmConfig,
-  llmConfig: LlmConnectionConfig & { supportsJsonSchema?: boolean },
+  llmConfig: LlmProfileConfig,
   chat: NonNullable<AkmExtractOptions["chat"]>,
   ctx: ProposalsContext | undefined,
   sourceRun: string,
@@ -843,7 +843,7 @@ export async function akmExtract(options: AkmExtractOptions): Promise<AkmExtract
   //      (per-process override, matches reflect/distill/consolidate)
   //   2. config.defaults.llm (the default LLM profile)
   //   3. throw — extract requires an LLM.
-  let llmConfig: (LlmConnectionConfig & { supportsJsonSchema?: boolean }) | undefined;
+  let llmConfig: LlmProfileConfig | undefined;
   const runnerSpec = resolveImproveProcessRunnerFromProfile(extractProcess, config);
   if (runnerSpec) {
     if (!runnerIsLlm(runnerSpec)) {
