+++ tests/commands/reflect-response-schema.test.ts
@@ -22,7 +22,7 @@ import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
 
-import type { LlmConnectionConfig } from "../../src/core/config/config";
+import type { LlmConnectionConfig, LlmProfileConfig } from "../../src/core/config/config";
 import { parseAgentProposalPayload } from "../../src/integrations/agent/prompts";
 
 // ── Module-level chatCompletion spy ─────────────────────────────────────────
@@ -84,7 +84,7 @@ function makeStashDir(): string {
   return stash;
 }
 
-function fakeLlmConnection(): LlmConnectionConfig & { supportsJsonSchema?: boolean } {
+function fakeLlmConnection(): LlmProfileConfig {
   return {
     endpoint: "http://localhost:11434/v1/chat/completions",
     model: "test-model",
